requirements.txt
altgraph==0.17.4
macholib==1.16.3
packaging==24.1
pillow==10.4.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
